CHG: Modify the minimal running parameters of DeepPrep to meet the requirements of BIDS-APP, and make them the same as fMRIPrep.
NEW: Generate the nextflow running configuration based on the command line command.

NEW: Use a shell script to unify the WorkDir of nextflow to the output_dir directory.

NEW: Add the functionality to set cpus and memory through the command line.

CHG: Optimize the input parameter quantity of the DeepPrep running command.